connectors/x-twitter/requirements.txt
tweepy==4.14.0
kafka-python==2.0.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
